requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
psycopg2==2.4.4
Flask-JSGlue
flask-jsonpify
